Add file project_referenced_libs.
This file is for recored the libs that project referenced. As we all know that eclipse store this information in file .classpath in the root directory of the project. But the file .classpath is add to file .gitignore, so when some one clone this project add import to eclipse they will need to add the libs that this project referenced.

This file is the place where people can find what external libs they need to add in order to satify the project.